+++ src/plugin/systems/writeback.rs
@@ -1,22 +1,30 @@
 use crate::dynamics::MassProperties;
 use crate::dynamics::ReadMassProperties;
+use crate::plugin::context::RapierContextEntityLink;
 use crate::plugin::RapierConfiguration;
-use crate::plugin::RapierContext;
+use crate::plugin::RapierContextAccess;
 use crate::prelude::MassModifiedEvent;
 use bevy::prelude::*;
 
 /// System responsible for writing updated mass properties back into the [`ReadMassProperties`] component.
 pub fn writeback_mass_properties(
-    mut context: ResMut<RapierContext>,
-    config: Res<RapierConfiguration>,
+    link: Query<&RapierContextEntityLink>,
+    context: RapierContextAccess,
+    config: Query<&RapierConfiguration>,
 
     mut mass_props: Query<&mut ReadMassProperties>,
     mut mass_modified: EventReader<MassModifiedEvent>,
 ) {
-    let context = &mut *context;
+    for entity in mass_modified.read() {
+        let link = link
+            .get(entity.0)
+            .expect("Could not find `RapierContextEntityLink`");
+        let config = config
+            .get(link.0)
+            .expect("Could not find `RapierConfiguration`");
+        if config.physics_pipeline_active {
+            let context = context.context(link);
 
-    if config.physics_pipeline_active {
-        for entity in mass_modified.read() {
             if let Some(handle) = context.entity2body.get(entity).copied() {
                 if let Some(rb) = context.bodies.get(handle) {
                     if let Ok(mut mass_props) = mass_props.get_mut(**entity) {